refactor hooks and states
