+++ src/components/page/ReceiverPage/hooks.ts
@@ -1,8 +1,10 @@
+import { useParams } from "next/navigation";
 import { useCallback, useRef } from "react";
 import { match } from "ts-pattern";
 
 import { useSockets } from "@/app/providers/socket";
-import { useTextMutation } from "@/usecase/mutation/hooks";
+import { useMutationStates } from "@/states";
+import { sendTextToAI } from "@/usecase";
 import { guardUndef } from "@/utils";
 
 export const useReceiver = () => {
@@ -10,9 +12,11 @@ export const useReceiver = () => {
     const {
         isMutating,
         mutatedDisplayIndex,
-        mutateText,
-        mutator: { finishMutation },
-    } = useTextMutation();
+        mutator: { startMutation, finishMutation },
+    } = useMutationStates();
+
+    const params = useParams();
+    const id = parseInt(params.id[0], 10);
 
     const receivedTextRef = useRef<HTMLDivElement>(null);
 
@@ -21,6 +25,26 @@ export const useReceiver = () => {
             setReceivedText(text);
         },
         [setReceivedText]
+    );
+
+    const mutateText = useCallback(
+        async (displayText: string, targetText: string) => {
+            startMutation();
+            const res = await sendTextToAI(targetText, id);
+            match(res)
+                .with({ status: "ok" }, () => {
+                    const mutatedText = guardUndef(res.val);
+                    console.log(mutatedDisplayIndex);
+                    const newDisplayText = displayText.slice(0, mutatedDisplayIndex) + mutatedText;
+                    updateText(newDisplayText);
+                    finishMutation(targetText, mutatedText);
+                })
+                .with({ status: "err" }, () => {
+                    console.log(res.err?.message);
+                    finishMutation(targetText, "");
+                });
+        },
+        [startMutation, updateText, finishMutation, id, mutatedDisplayIndex]
     );
 
     const handleInputChange = useCallback(async () => {
@@ -32,20 +56,10 @@ export const useReceiver = () => {
 
         // 5回以上の場合は mutation 実行
         if (count >= 5 && !isMutating) {
-            console.log("句読点または改行が5回以上入力されました。");
-            const res = await mutateText(targetText);
-            match(res)
-                .with({ status: "ok" }, () => {
-                    const mutatedText = guardUndef(res.val);
-                    const displayText = text.slice(0, mutatedDisplayIndex) + mutatedText;
-                    updateText(displayText);
-                    finishMutation(targetText, mutatedText);
-                })
-                .with({ status: "err" }, () => {
-                    console.log(res.err?.message);
-                });
+            console.log(`句点または改行が5回以上入力されました。: ${targetText}`);
+            await mutateText(text, targetText);
         }
-    }, [updateText, isMutating, mutateText, finishMutation, mutatedDisplayIndex]);
+    }, [isMutating, mutateText, mutatedDisplayIndex]);
 
     return {
         receivedTextRef,
