+++ src/components/page/StreamerPage/hooks.ts
@@ -1,7 +1,6 @@
 import React, { useCallback } from "react";
 
 import { useSockets } from "@/app/providers/socket";
-
 import { useStreamService } from "@/usecase";
 
 export const useStream = () => {
