--- src/usecase/mutation/index.ts
@@ -1,2 +1,0 @@
-export { useTextMutation } from "./hooks";
-export { sendTextToAI } from "./service";
