--- src/usecase/streaming/index.ts
@@ -1,2 +1,0 @@
-export { useStreamService } from "./streamer";
-export { useReceiveService } from "./receiver";
